feat: Network rescanning, signal strength, better network name parsing
The first version of "wifi" assumed that there are no spaces in network
names.

This naive assumption did not work when I wanted to connect to my phone
hotspot, which contains whitespaces.

Therefore, the network parsing logic is changed to get the correct
network name.

Also, the output of scan is improved to show users the signal strength
of each network.

Last but not least, there might be occasions where the initial network
scan may not pick up the desired network. For these situations,
rescanning is added to the scan command.

Rescan triggers when users type "scan" to the prompt where they see a
list of scanned networks.